+++ cmake/TestseT.cmake
@@ -76,8 +76,9 @@ add_eT_runtest(hf_oxonium_charge_sad                        "eT;short;hf;sad;cha
 add_eT_runtest(restart_hf                                   "eT;short;hf;restart")
 #
 add_eT_runtest(hf_visualization                             "eT;short;hf;visualization")
-add_eT_runtest(plot_benzene_mos                             "eT;short;hf;visualization")
+add_eT_runtest(plot_h2o_cartesian_mos                       "eT;short;hf;visualization,cartesian")
 add_eT_runtest(plot_He_orbitals                             "eT;short;hf;visualization;5z")
+add_eT_runtest(plot_benzene_mos                             "eT;short;hf;visualization")
 add_eT_runtest(plot_ntos_ccs                                "eT;short;ccs;visualization")
 add_eT_runtest(plot_ntos_ccsd                               "eT;short;ccsd;visualization")
 add_eT_runtest(plot_cntos_ccsd                              "eT;short;ccsd;visualization")