+++ src/various/angular_momentum/cartesian_f_angular_momentum.F90
@@ -29,10 +29,10 @@ module cartesian_f_angular_momentum_class
 !!    https://github.com/evaleev/libint/wiki/using-modern-CPlusPlus-API
 !!
 !
-   use abstract_angular_momentum_class
+   use abstract_cartesian_angular_momentum_class
 !
 !
-   type, extends(abstract_angular_momentum) :: cartesian_f_angular_momentum
+   type, extends(abstract_cartesian_angular_momentum) :: cartesian_f_angular_momentum
    end type
 !
    interface cartesian_f_angular_momentum
@@ -58,9 +58,11 @@ contains
 !
       this%l_letter = 'f'
 !
-      allocate(this%offsets(10))
-      allocate(this%components(10))
-      allocate(this%normalization(10))
+      this%n_functions = 10
+!
+      allocate(this%offsets(this%n_functions))
+      allocate(this%components(this%n_functions))
+      allocate(this%normalization(this%n_functions))
 !
 !     instead of   xxx, xxy, xxz, xyy, xyz, xzz, yyy, yyz, yzz, zzz
 !     Molden needs xxx, yyy, zzz, xyy, xxy, xxz, xzz, yzz, yyz, xyz