+++ src/various/angular_momentum/cartesian_h_angular_momentum.F90
@@ -29,10 +29,10 @@ module cartesian_h_angular_momentum_class
 !!    https://github.com/evaleev/libint/wiki/using-modern-CPlusPlus-API
 !!
 !
-   use abstract_angular_momentum_class
+   use abstract_cartesian_angular_momentum_class
 !
 !
-   type, extends(abstract_angular_momentum) :: cartesian_h_angular_momentum
+   type, extends(abstract_cartesian_angular_momentum) :: cartesian_h_angular_momentum
    end type
 !
    interface cartesian_h_angular_momentum
@@ -60,9 +60,11 @@ contains
 !
       this%l_letter = 'h'
 !
-      allocate(this%offsets(21))
-      allocate(this%components(21))
-      allocate(this%normalization(21))
+      this%n_functions = 21
+!
+      allocate(this%offsets(this%n_functions))
+      allocate(this%components(this%n_functions))
+      allocate(this%normalization(this%n_functions))
 !
 !     Molden cannot handle l > 4
       this%offsets = [(i, i=1, 21)]