+++ src/various/angular_momentum/spherical_i_angular_momentum.F90
@@ -29,9 +29,9 @@ module spherical_i_angular_momentum_class
 !!    https://github.com/evaleev/libint/wiki/using-modern-CPlusPlus-API
 !!
 !
-   use abstract_angular_momentum_class
+   use abstract_spherical_angular_momentum_class
 !
-   type, extends(abstract_angular_momentum) :: spherical_i_angular_momentum
+   type, extends(abstract_spherical_angular_momentum) :: spherical_i_angular_momentum
    end type
 !
 !
@@ -60,9 +60,11 @@ contains
 !
       this%l_letter = 'i'
 !
-      allocate(this%offsets(13))
-      allocate(this%components(13))
-      allocate(this%normalization(13))
+      this%n_functions = 13
+!
+      allocate(this%offsets(this%n_functions))
+      allocate(this%components(this%n_functions))
+      allocate(this%normalization(this%n_functions))
 !
 !     Molden cannot handle l > 4
       this%offsets = [(i, i=1, 13)]