+++ changelog.md
@@ -7,7 +7,7 @@
 - Added unit tests for for untested routines in `tools/maps`. eT-program/eT!1035
 
 ### Structure
-- Cleanup of engines, by introducing tasks that are called from the engine. eT-program/eT!968, eT-program/eT!1040
+- Cleanup of engines, by introducing tasks that are called from the engine. eT-program/eT!968, eT-program/eT!1040, eT-program/eT!1044
 
 # eT v1.6.1
 ### README